simplification of the API:
engine is now initialised with the origin URL rather that origin URL PLUS worklet name. worklet name is now derived from URL. this has to be well documented

samples are now loaded with a relative URL to the origin URL that the engine was initialized with.

learners creation, initialisation and comms with engine have been simplified. Learners are now injected to the engine with `addLearner` ASYNC method which:
1. initialises the Learner with the origin URL
2. sets up up async comms between engine and Learner
3. add Learner to learners hash